exec: cleanup
Create*AsCurrentUser now takes an 'interactive' parameter to decide whether
the process should be launched in the active console session.
CreatePipedProcessAsUser: user can be NULL (use current).
CreatePipedProcessAsUser: using a user different than a currently logged on
one returns an error. This is broken and needs fixing.
GrantRemoteSessionDesktopAccess needs to give desktop/winsta access to such
user, but it's not possible without running a new process in the interactive
session (we're usually running in session 0 as a service). This was originally
implemented with a hack that launched qrexec-agent with a special parameter,
this needs a proper solution (another executable).